Revemoing generics exceptions and filtering by User

+++ tasks-service/src/main/java/tech/engix/tasks_service/service/TasksService.java
@@ -4,13 +4,17 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import tech.engix.jwtutils.service.JwtUtils;
+import tech.engix.tasks_service.client.ProjectServiceClient;
 import tech.engix.tasks_service.client.UserServiceClient;
+import tech.engix.tasks_service.dto.ProjectClientResponse;
 import tech.engix.tasks_service.dto.RequestTasks;
 import tech.engix.tasks_service.dto.TasksUpdateRequest;
 import tech.engix.tasks_service.dto.UserResponse;
 import tech.engix.tasks_service.mapper.TasksMapper;
 import tech.engix.tasks_service.model.Tasks;
 import tech.engix.tasks_service.repository.TaskRepository;
+import tech.engix.tasks_service.service.exception.exceptions.AccessDeniedException;
+import tech.engix.tasks_service.service.exception.exceptions.ClientNotFound;
 
 import java.time.LocalDateTime;
 import java.util.Collections;
@@ -29,7 +33,9 @@ public class TasksService {
 
     private final UserServiceClient userServiceClient;
 
-    public List<Tasks> listAll(String jwtToken) {
+    private final ProjectServiceClient projectServiceClient;
+
+    public List<Tasks> listAllByUser(String jwtToken) {
         String email = jwtUtils.getUserNameFromJwtToken(jwtToken);
 
         UserResponse user = userServiceClient.getUserByEmail(email);
@@ -41,13 +47,28 @@ public class TasksService {
         return Collections.emptyList();
     }
 
+    public List<Tasks> listTasksByProject(Long projectId, String jwtToken) {
+        String email = jwtUtils.getUserNameFromJwtToken(jwtToken);
+        UserResponse user = userServiceClient.getUserByEmail(email);
+        if (repository.existsByProjectIdAndUserId(projectId, user.id())) {
+            return repository.findByProjectId(projectId);
+        } else {
+            throw new ClientNotFound("Client not found or user not authorized to delete this client.");
+        }
+    }
 
-    public RequestTasks createTasks(RequestTasks requestTasks, String jwtToken) {
+    public RequestTasks createTasks(RequestTasks requestTasks, String jwtToken, Long id) {
         String email = jwtUtils.getUserNameFromJwtToken(jwtToken);
         UserResponse user = userServiceClient.getUserByEmail(email);
 
+        ProjectClientResponse project = projectServiceClient.getClientName(id);
+
         Tasks tasks = mapper.toDto(requestTasks);
+
         tasks.setUserId(user.id());
+        tasks.setProjectId(project.id());
+        tasks.setProjectName(project.name());
+
         tasks.setCreatedAt(LocalDateTime.now());
         repository.save(tasks);
 
@@ -70,14 +91,22 @@ public class TasksService {
                 tasks.setCreatedAt(LocalDateTime.now());
                 return mapper.toUpdate(updatedClient);
             } else {
-                throw new RuntimeException("Unauthorized: You do not have permission to edit this tasks.");
+                throw new AccessDeniedException("Unauthorized: You do not have permission to edit this tasks.");
             }
         } else {
-            throw new RuntimeException("Client not found with id " + id);
+            throw new ClientNotFound("Client not found with id " + id);
         }
     }
 
-    public void deleteTasks(String id) {
-        repository.deleteById(id);
+
+    public void deleteTasks(String id, String jwtToken) {
+        String email = jwtUtils.getUserNameFromJwtToken(jwtToken);
+        UserResponse user = userServiceClient.getUserByEmail(email);
+
+        if (repository.existsByIdAndUserId(id, user.id())) {
+            repository.deleteById(id);
+        } else {
+            throw new ClientNotFound("Client not found or user not authorized to delete this client.");
+        }
     }
 }
